refactor: Consolidate common code and fix bugs
This commit refactors the project by consolidating common code into the `core/common` module and addresses a few bugs.

**Key changes:**

-   **Code Consolidation**:
    -   Removed boilerplate `Platform.kt` and its platform-specific implementations from all modules (`composeApp`, `core/navigation`, `core/network`, `feature/dashboard`, `feature/splashscreen`).
    -   Centralized the `Platform` interface and `getPlatform()` implementations within the `core/common` module.
    -   Moved all DataStore-related files (`createDataStore`, `DataStoreHelper`, `ContextUtils`, etc.) to a more consistent package structure under `my.id.tasius.common`.

-   **Bug Fixes**:
    -   Corrected a logic error in `SplashScreenRepositoryImpl` where the `IS_FIRST_RUN_LOCAL_KEY` was incorrectly set, causing it to always return `true`.
    -   Fixed a navigation issue by updating `navigateAndClearBackStack` to use `graph.findStartDestination().id` for more reliable back stack clearing.
    -   In `DataStoreHelperImpl`, the `getBoolean` method now correctly returns `null` if the key does not exist, instead of defaulting to `false`.

-   **Project Structure**:
    -   Updated package names and import statements across the app to reflect the refactored module structure (e.g., `my.id.tasius.dailysleeptracker.core.common` to `my.id.tasius.common`).
    -   Cleaned up the `Greeting.kt` boilerplate code from the initial project setup.
