+++ composeApp/src/androidMain/kotlin/my/id/tasius/dailysleeptracker/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import my.id.tasius.dailysleeptracker.core.common.storage.ContextUtils
+import my.id.tasius.common.storage.ContextUtils
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
