--- composeApp/src/androidMain/kotlin/my/id/tasius/dailysleeptracker/Platform.android.kt
@@ -1,9 +1,0 @@
-package my.id.tasius.dailysleeptracker
-
-import android.os.Build
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
