--- composeApp/src/commonMain/kotlin/my/id/tasius/dailysleeptracker/Greeting.kt
@@ -1,9 +1,0 @@
-package my.id.tasius.dailysleeptracker
-
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
