--- composeApp/src/commonMain/kotlin/my/id/tasius/dailysleeptracker/Platform.kt
@@ -1,7 +1,0 @@
-package my.id.tasius.dailysleeptracker
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
