--- composeApp/src/iosMain/kotlin/my/id/tasius/dailysleeptracker/Platform.ios.kt
@@ -1,9 +1,0 @@
-package my.id.tasius.dailysleeptracker
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
