+++ core/common/src/androidMain/kotlin/my/id/tasius/common/Platform.android.kt
@@ -1,3 +1,9 @@
 package my.id.tasius.common
 
-actual fun platform() = "Android"
+import android.os.Build
+
+class AndroidPlatform : Platform {
+    override val name: String = "Android ${Build.VERSION.SDK_INT}"
+}
+
+actual fun getPlatform():Platform = AndroidPlatform()
