+++ core/common/src/androidMain/kotlin/my/id/tasius/common/storage/ContextUtils.kt
@@ -1,0 +1,12 @@
+package my.id.tasius.common.storage
+
+import android.content.Context
+
+object ContextUtils {
+
+    var dataStoreApplicationContext: Context? = null
+
+    fun setContext(context: Context) {
+        dataStoreApplicationContext = context.applicationContext
+    }
+}
