+++ core/common/src/androidMain/kotlin/my/id/tasius/common/storage/createDataStore.android.kt
@@ -1,0 +1,15 @@
+package my.id.tasius.common.storage
+
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import kotlinx.coroutines.runBlocking
+
+actual fun createDataStore(): DataStore<Preferences> {
+    return runBlocking {
+        getDataStore(producePath = {
+            ContextUtils.dataStoreApplicationContext?.filesDir?.resolve(
+                dataStoreFileName
+            )?.absolutePath ?: ""
+        })
+    }
+}
