--- core/common/src/androidMain/kotlin/my/id/tasius/dailysleeptracker/core/common/storage/ContextUtils.kt
@@ -1,12 +1,0 @@
-package my.id.tasius.dailysleeptracker.core.common.storage
-
-import android.content.Context
-
-object ContextUtils {
-
-    var dataStoreApplicationContext: Context? = null
-
-    fun setContext(context: Context) {
-        dataStoreApplicationContext = context.applicationContext
-    }
-}
