+++ core/common/src/commonMain/kotlin/my/id/tasius/common/Platform.kt
@@ -1,3 +1,7 @@
 package my.id.tasius.common
 
-expect fun platform(): String
+interface Platform {
+    val name: String
+}
+
+expect fun getPlatform(): Platform
