--- core/common/src/commonMain/kotlin/my/id/tasius/dailysleeptracker/core/common/storage/createDataStore.kt
@@ -1,28 +1,0 @@
-package my.id.tasius.dailysleeptracker.core.common.storage
-
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.PreferenceDataStoreFactory
-import androidx.datastore.preferences.core.Preferences
-import kotlinx.coroutines.InternalCoroutinesApi
-import kotlinx.coroutines.internal.SynchronizedObject
-import kotlinx.coroutines.internal.synchronized
-import okio.Path.Companion.toPath
-
-expect fun createDataStore(): DataStore<Preferences>
-
-internal const val dataStoreFileName = "create_datastore.preferences_pb"
-
-private lateinit var dataStore: DataStore<Preferences>
-@OptIn(InternalCoroutinesApi::class)
-private val lock = SynchronizedObject()
-
-@OptIn(InternalCoroutinesApi::class)
-fun getDataStore(producePath: () -> String): DataStore<Preferences> =
-    synchronized(lock) {
-        if (::dataStore.isInitialized) {
-            dataStore
-        } else {
-            PreferenceDataStoreFactory.createWithPath(produceFile = { producePath().toPath() })
-                .also { dataStore = it }
-        }
-    }
