+++ core/common/src/iosMain/kotlin/my/id/tasius/common/Platform.ios.kt
@@ -1,3 +1,9 @@
 package my.id.tasius.common
 
-actual fun platform() = "iOS"
+import platform.UIKit.UIDevice
+
+class IOSPlatform: Platform {
+    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
+}
+
+actual fun getPlatform(): Platform = IOSPlatform()
