--- core/common/src/iosMain/kotlin/my/id/tasius/dailysleeptracker/core/common/storage/createDataStore.ios.kt
@@ -1,26 +1,0 @@
-package my.id.tasius.dailysleeptracker.core.common.storage
-
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import kotlinx.cinterop.ExperimentalForeignApi
-import kotlinx.coroutines.runBlocking
-import platform.Foundation.NSDocumentDirectory
-import platform.Foundation.NSFileManager
-import platform.Foundation.NSURL
-import platform.Foundation.NSUserDomainMask
-
-@OptIn(ExperimentalForeignApi::class)
-actual fun createDataStore(): DataStore<Preferences> {
-    return runBlocking {
-        getDataStore(producePath = {
-            val documentDirectory: NSURL? = NSFileManager.defaultManager.URLForDirectory(
-                directory = NSDocumentDirectory,
-                inDomain = NSUserDomainMask,
-                appropriateForURL = null,
-                create = false,
-                error = null,
-            )
-            requireNotNull(documentDirectory).path + "/${dataStoreFileName}"
-        })
-    }
-}
