--- core/navigation/src/androidMain/kotlin/my/id/tasius/navigation/Platform.android.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.navigation
-
-actual fun platform() = "Android"
