--- core/navigation/src/commonMain/kotlin/my/id/tasius/navigation/Platform.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.navigation
-
-expect fun platform(): String
