--- core/navigation/src/iosMain/kotlin/my/id/tasius/navigation/Platform.ios.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.navigation
-
-actual fun platform() = "iOS"
