--- core/network/src/androidMain/kotlin/my/id/tasius/network/Platform.android.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.network
-
-actual fun platform() = "Android"
