--- core/network/src/commonMain/kotlin/my/id/tasius/network/Platform.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.network
-
-expect fun platform(): String
