--- core/network/src/iosMain/kotlin/my/id/tasius/network/Platform.ios.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.network
-
-actual fun platform() = "iOS"
