--- feature/dashboard/src/androidMain/kotlin/my/id/tasius/dashboard/Platform.android.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.dashboard
-
-actual fun platform() = "Android"
