--- feature/dashboard/src/commonMain/kotlin/my/id/tasius/dashboard/Platform.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.dashboard
-
-expect fun platform(): String
