--- feature/dashboard/src/iosMain/kotlin/my/id/tasius/dashboard/Platform.ios.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.dashboard
-
-actual fun platform() = "iOS"
