--- feature/splashscreen/src/androidMain/kotlin/my/id/tasius/splashscreen/Platform.android.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.splashscreen
-
-actual fun platform() = "Android"
