--- feature/splashscreen/src/commonMain/kotlin/my/id/tasius/splashscreen/Platform.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.splashscreen
-
-expect fun platform(): String
