--- feature/splashscreen/src/iosMain/kotlin/my/id/tasius/splashscreen/Platform.ios.kt
@@ -1,3 +1,0 @@
-package my.id.tasius.splashscreen
-
-actual fun platform() = "iOS"
